start abstracting chromedriver stuff out of chrome.js so we can reuse it for webviews
